FINAL/IllusoryBeacon/backend/requirements.txt
aiofiles==0.4.0
aiosqlite==0.11.0
asyncpg==0.20.0
certifi==2019.11.28
chardet==3.0.4
ciso8601==2.1.2
dnspython==1.16.0
docutils==0.15.2
email-validator==1.0.5
geohash2==1.1
h11==0.8.1
h2==3.1.1
hpack==3.0.0
httpcore==0.3.0
httptools==0.0.13
hyperframe==5.2.0
idna==2.8
multidict==4.7.1
PyPika==0.35.16
requests==2.22.0
requests-async==0.5.0
rfc3986==1.3.2
sanic==19.9.0
Sanic-Auth==0.2.0
sanic-session==0.5.6
tortoise-orm==0.15.5
typing-extensions==3.7.4.1
ujson==1.35
urllib3==1.25.7
uvloop==0.14.0
websockets==8.1